Build AppImage that could start a Terminal Vim
The AppImage now doesn't start GVim directly, but calls a script
to start Vim. This script evaluates the name that was used to start
the AppImage. If the name starts with 'vim', the terminal version is
started, else GVim is started.

Prerequisite for this was the use of AppIMageKit V11 (newest release).

Also added the script 'scripts/build_vim.sh' that is used to configure
and make Vim. This script is also used from '.travis.yml'.