release: provide a nicer release log
Currently the Release page is a bit terse. This commit makes it a bit
more nicer for the users:
- it directly links the artifacts for download
- it mentions what artifact to use for what use case
- it mentions how to run it
- it directly links to the Vim Commits since the last change

The release page will be generated by the scripts/release_notes.sh
script, but since it needs access to the various environment properties
of the appimage.sh script, it will not be called, but rather be source,
that's why it's using the '. release_notes.sh' syntax, so that it is
executed in the current shell environment.

Also, while at it, slightly refactor the Github Actions release
workflow and rename the jobs with the `-job` prefix, so that there does
no longer exist a `check-updates` job as well as `check-updates` step,
which is confusing if you need to reference the output variable.